Corrected previous commit.
  The previous commit was causing a crash because a change to the UI
thread wasn't run inside Androids special runOnUiThread method. This
error got passed me because I had tested an earlier version where I
changed the main UI late when showing a banner ad, instead of on
initializing the plugin, and for some reason the version that doesn't
force the setLayout call to run on the UI thread works in my app. :)